closes #277 by way of adding an if clause within clear simdata function that checks to see if worker is initialised before attempting to terminate it
Former-commit-id: 065c287598a6faa8edeac36a1524b0a213cae3f9

+++ RedR-Disaster-Simulator/DisasterSimulatorNodeJSServer/server.js
@@ -120,7 +120,9 @@ function clearSimData() {
     };
     connectedUsers = [];
     connectedUsers.push(host);
-    worker.terminate();
+    if(worker != null) {
+        worker.terminate();
+    }
 
 
 }
